UI elements and player Inventory
Added elements for UI such as player HUD (health and mana sliders, three hotkey buttons, map and inventory buttons), temp versions of Inventory menu and map menu.
Made classes for Inventory menu and map menu, with a UI manager, and made it such that the inventory and map pop up on clicking the corresponding buttons (though it only works after a first click, for some reason).
Made classes for Inventory management; specifically, a bunch of enum classes (ItemName, ItemEffect, SpellName, SpellEffect) as well as Item and Spell classes, the lot of which is combined in the PlayerInventory class. It's a good foundation to build upon.