Camera motion and spell casting animation state
Added camera motion in accordance with player input (really, I just added the camera as a child of the player game object, like Sabin suggested; this way with every movement the player makes, the camera moves too. We can even rotate, if we need to (which we shouldn't)).
Added a transition from the Default Movement Template animator controller to the CastSpell animator state (which, as of yet, is just some little temp animation of a magic wand sprite, though you won't be able to distinguish the wand (it's there, though, trust.)). The player movement state machine has now been compressed to a single little node, so things will be easier to manage from there. See the player animator for more information.