Enemy and rough wandering AI
Implemented an enemy gameobject, added an EnemyMovement script responsible for its movement wherein we update the enemy's animator parameters (that have to do with movement). The enemy moves around inside a given region that will serve as a detection area. The premise is that the enemy is effectively (randomly) patrolling this region and is bound to it.