MobileCharacter modifications
I added a [m_Direction] variable to all mobile character classes inheriting from MobileCharacter, reasoning that since there won't ever be any rotation of any of the gameobjects, this would be the quicker way to see which way people are facing.
This is mainly a quick update before I start working on other things.