Fixed Spellcasting
Implemented a sort of default spell-casting wherein a spell will be cast even if there is no target; it will simply go in the direction in which the player clicked on the screen. Inversely, if the player clicks on an enemy, then the spell will target and lock-on to the enemy.
The spell is destroyed after two seconds, and if the spell reaches its enemy target it is destroyed.