Obstructable
Implemented collision with scenery objects; just throw on the Obstructable.cs script on whatever you want the player to not be able to move through.
Note: You'll need to set your own collider such that the player not pass through it, properly. Don't forget to account for the player's slantedness (this brings him a little closer to some objects than others, particularly tall ones).