Spell pickup fix and bug report
Okay so the last upload concerning pickups wasn't working properly; I verified this after creating new spells and items to check my work. Good call, that was. I fixed it for the spells, but I couldn't fix it for items.
What's happening in the cases of both List and Dictionary in the player inventory is that lists can't tell spells apart and dictionaries can't tell items apart. It sucks. I tried to fix it manually with our own equals(other) method and it worked for the lists, but not for the dictionary :/
The rest of the scene is unchanged and works perfectly. Just don't hit 1, 2, 3, or 4; otherwise you'll spawn stuff in.