Implemented Thunderball & Bug
Okay so Thunderball is completely implemented but I also discovered a pretty substantial bug: the spells, for some reason, are triggering the OnTriggerEnter function twice on the enemies, doubling the effect of any given spell.
I don't know why this is happening, I've been trying to solve it for a few hours now, and still nothing. Gonna move onto something else in the meantime.