Various
Implemented death conditions in animator and script - only need an animation on the player states.
Tried to find a way to serialize the player data for saving and loading; didn't manage to serialize it but I created a class that should be serializable, called Serializable_Player. This class has everything we need to save and load the game, we just need to figure out where to put it.
Implemented icons/sprites for items.