Position fix
Right, so here's what the problem was: when you spawn in the player, for a fraction of a second, their health is 0. Because of this the Resurrect function was being called. So I set the player's respawn position to be the position we saved. What this means is that it's likely that a Game Over will be triggered when we load in the scene, as things are. And that reminds me, we need a game over screen.