Rough scene transition from demo scene to overworld
Dunno why but the player's position keeps being stubborn.
I've tried to fix it in the Player class, at the bottom of the Start function. It works when we're moving into the overworld scene, but for some reason we keep getting spawned at that initial [m_PlayerRespawnLocation] when we're going from overworld to demo. Still, we have successful scene transition!